requirements.txt
paypalrestsdk~=1.13.1
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
sendgrid~=6.8.3
python-dotenv
functions-framework
pyyaml
